pyproject.toml
[project]
name = "check-cvs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfreader>=0.1.15",
    "pypdf2<3.0.0",
    "python-docx>=1.2.0",
]
